Add the compiler module to docs
