+++ doc/code/qml_compiler.rst
@@ -1,0 +1,8 @@
+qml.compiler
+===============
+
+.. currentmodule:: pennylane.compiler
+
+.. automodapi:: pennylane.compiler
+    :no-heading:
+    :include-all-objects:
