+++ doc/index.rst
@@ -183,6 +183,7 @@ PennyLane is **free** and **open source**, released under the Apache License, Ve
    :hidden:
 
    code/qml
+   code/qml_compiler
    code/qml_data
    code/qml_drawer
    code/qml_fermi
